made changes suggested by Rashnil --> renaming specific files
