+++ v2/get_data.py
@@ -5,8 +5,8 @@ import boto3
 import botocore
 
 # Import custom stuff
-from standard_functions.datetime import find_date_in_str
-from standard_functions.get_creds import get_aws_creds
+from utils.datetime import find_date_in_str
+from utils.get_creds import get_aws_creds
 
 
 # Credentials
