+++ v2/main.py
@@ -12,8 +12,8 @@ pd.options.display.max_columns = 100
 
 
 # Load custom functions
-from preprocess_data import get_preprocessed_df
-from standard_functions.split_to_train_test import split_and_balance_with_SMOTE
+from user_reports_preprocessor import get_preprocessed_df
+from utils.split_to_train_test import split_and_balance_with_SMOTE
 
 
 # ==============================================================================
